+++ AEWatchRenderManager/MainWindow.xaml.cs
@@ -18,8 +18,8 @@ using System.Windows.Controls;
 
 namespace AEWatchRenderManager
 {
-    // Date: Tue Apr 14 12:47:38 JST 2026
-    // Version: 1.16.11
+    // Date: Wed Apr 15 11:02:12 JST 2026
+    // Version: 1.16.12
     public partial class MainWindow : Window
     {
         public MainWindow()