+++ AEWatchRenderManager/Models/RenderTaskPair.cs
@@ -5,8 +5,8 @@ using System.IO;
 
 namespace AEWatchRenderManager.Models
 {
-    // Date: Tue Apr 14 12:47:38 JST 2026
-    // Version: 1.16.11
+    // Date: Wed Apr 15 11:02:12 JST 2026
+    // Version: 1.16.12
     public enum RenderStatus
     {
         Queued,    // 待機中