+++ AEWatchRenderManager/Services/TaskPairManager.cs
@@ -9,8 +9,8 @@ using System.Threading.Tasks;
 
 namespace AEWatchRenderManager.Services
 {
-    // Date: Tue Apr 14 12:47:38 JST 2026
-    // Version: 1.16.11
+    // Date: Wed Apr 15 11:02:12 JST 2026
+    // Version: 1.16.12
     public class TaskPairManager
     {
         public ObservableCollection<RenderTaskPair> Tasks { get; } = new();